+++ src/__generated__/Show_show.graphql.ts
@@ -1,6 +1,7 @@
 /* tslint:disable */
 
 import { ConcreteFragment } from "relay-runtime";
+import { Artworks_show$ref } from "./Artworks_show.graphql";
 import { ShowHeader_show$ref } from "./ShowHeader_show.graphql";
 declare const _Show_show$ref: unique symbol;
 export type Show_show$ref = typeof _Show_show$ref;
@@ -28,18 +29,7 @@ export type Show_show = {
         readonly name: string | null;
         readonly is_followed: boolean | null;
     }) | null> | null;
-    readonly artworks: ReadonlyArray<({
-        readonly __id: string;
-        readonly id: string;
-        readonly artist_names: string | null;
-        readonly image: ({
-            readonly id: string | null;
-            readonly url: string | null;
-        }) | null;
-        readonly price: string | null;
-        readonly availability: string | null;
-        readonly contact_label: string | null;
-    }) | null> | null;
+    readonly status: string | null;
     readonly counts: ({
         readonly artworks: number | null;
         readonly eligible_artworks: any | null;
@@ -47,7 +37,7 @@ export type Show_show = {
     readonly partner: ({
         readonly name?: string | null;
     }) | null;
-    readonly " $fragmentRefs": ShowHeader_show$ref;
+    readonly " $fragmentRefs": ShowHeader_show$ref & Artworks_show$ref;
     readonly " $refType": Show_show$ref;
 };
 
@@ -204,64 +194,16 @@ return {
       ]
     },
     {
-      "kind": "LinkedField",
+      "kind": "FragmentSpread",
+      "name": "Artworks_show",
+      "args": null
+    },
+    {
+      "kind": "ScalarField",
       "alias": null,
-      "name": "artworks",
-      "storageKey": null,
+      "name": "status",
       "args": null,
-      "concreteType": "Artwork",
-      "plural": true,
-      "selections": [
-        v1,
-        v0,
-        {
-          "kind": "ScalarField",
-          "alias": null,
-          "name": "artist_names",
-          "args": null,
-          "storageKey": null
-        },
-        {
-          "kind": "LinkedField",
-          "alias": null,
-          "name": "image",
-          "storageKey": null,
-          "args": null,
-          "concreteType": "Image",
-          "plural": false,
-          "selections": [
-            v0,
-            {
-              "kind": "ScalarField",
-              "alias": null,
-              "name": "url",
-              "args": null,
-              "storageKey": null
-            }
-          ]
-        },
-        {
-          "kind": "ScalarField",
-          "alias": null,
-          "name": "price",
-          "args": null,
-          "storageKey": null
-        },
-        {
-          "kind": "ScalarField",
-          "alias": null,
-          "name": "availability",
-          "args": null,
-          "storageKey": null
-        },
-        {
-          "kind": "ScalarField",
-          "alias": null,
-          "name": "contact_label",
-          "args": null,
-          "storageKey": null
-        }
-      ]
+      "storageKey": null
     },
     {
       "kind": "LinkedField",
@@ -314,5 +256,5 @@ return {
   ]
 };
 })();
-(node as any).hash = '3f7c6962f39801055cb6df605c2931bd';
+(node as any).hash = '35c6537d4bc579f5732e846891b81a87';
 export default node;