+++ src/lib/Scenes/Show/Components/Artworks/__tests__/index-tests.tsx
@@ -3,7 +3,12 @@ import "react-native"
 import * as renderer from "react-test-renderer"
 import { Artworks } from "../index"
 
-it("looks correct when rendererd", () => {
-  const comp = renderer.create(<Artworks artworks={[]} />)
+it("looks correct when rendered", () => {
+  const data = {
+    artworks: [],
+    __id: "12345",
+    $refType: {},
+  }
+  const comp = renderer.create(<Artworks show={data as any} />)
   expect(comp).toMatchSnapshot()
 })