+++ src/lib/Scenes/Show/Components/Shows/__tests__/index-tests.tsx
@@ -3,7 +3,7 @@ import "react-native"
 import * as renderer from "react-test-renderer"
 import { Shows } from "../index"
 
-it("looks correct when rendererd", () => {
+it("looks correct when rendered", () => {
   const comp = renderer.create(<Shows shows={[]} />)
   expect(comp).toMatchSnapshot()
 })